+++ src/overrides.js
@@ -1,5 +1,5 @@
 import {moduleName, midiQOL} from "../sharedvision.js";
-import {getPermission} from "./misc.js";
+import {getPermission, isSharedVision} from "./misc.js";
 
 let old_isVisionSource = Token.prototype._isVisionSource;
 
@@ -8,35 +8,7 @@ export function isVisionSourceOverride() {
     if (midiQOL) result = isVisionSourceOverride_MidiQOLFix.call(this);
     else result = old_isVisionSource.call(this);
     if (result) return true;
-  
-    let sharedVision = false;
-    if (game.user.isGM == false && this.actor != null) {
-        if (this.data.hidden) {
-            if ( (midiQOL && getPermission(this.actor,game.user, "OWNER")) == false && this.actor.data.flags.SharedVision?.hidden == false) return false;
-        }
-  
-        if (game.settings.get(moduleName,'enable')) {
-            sharedVision = this.actor.data.flags.SharedVision != undefined ? this.actor.data.flags.SharedVision.enable : false; 
-        }
-
-        if (sharedVision == false) {
-            const userSetting = this.actor.data.flags.SharedVision?.userSetting;
-            if (userSetting != undefined) {
-                for (let setting of userSetting)
-                    if (setting.id == game.userId) {
-                        sharedVision = setting.enable;
-                        break;
-                    }
-            }
-        }
-        
-        if (sharedVision == false) {
-            const permission = this.actor.data.permission?.[game.userId] ? this.actor.data.permission?.[game.userId] : this.actor.data.permission.default;
-            sharedVision = (permission==0 && game.settings.get(moduleName,'none')) || (permission==1 && game.settings.get(moduleName,'limited')) || (permission==2 && game.settings.get(moduleName,'observer')) || (permission==3 && game.settings.get(moduleName,'owner'));
-        }
-        return sharedVision;
-    } 
-    return result;
+    return isSharedVision(this);
 }
   
 //Copied from Midi QOL's patching.js
@@ -70,4 +42,30 @@ function isVisionSourceOverride_MidiQOLFix() {
 
     //TP ** const others = this.layer.controlled.filter(t => !t.data.hidden && t.hasSight);
     return !others.length;
+  }
+
+
+//Override of ForegroundLayer.prototype.updateOcclusion
+export function updateOcclusionOverride() {
+    let tokens = [];
+
+    if (game.user.isGM) tokens = canvas.tokens.controlled;
+    else {
+        const allTokens = canvas.tokens.children[0].children;
+        for (let token of allTokens) 
+            if (token.isOwner || isSharedVision(token)) 
+                tokens.push(token);
+    }
+
+    this._drawOcclusionShapes(tokens);
+    this.occlusionMask.roofs.removeChildren();
+    for ( let tile of this.tiles ) {
+      tile.updateOcclusion(tokens);
+      if ( tile.isRoof && (tile.occluded || !this.displayRoofs) ) {
+        const s = tile.getRoofSprite();
+        if ( !s ) continue;
+        s.tint = 0x0000FF;
+        this.occlusionMask.roofs.addChild(s);
+      }
+    }
   }